+++ src/spritesheet.cpp
@@ -49,7 +49,7 @@ Spritesheet::Spritesheet(Gui *gui, const ByteBuffer &key) :
     rucksack_texture_get_images(texture, images.raw());
     float full_width = tex_image._width; 
     float full_height = tex_image._height;
-    for (long i = 0; i < images.length(); i += 1) {
+    for (int i = 0; i < images.length(); i += 1) {
         RuckSackImage *image = images.at(i);
 
         SpritesheetImage *img = create<SpritesheetImage>();
